test: popover 테스트 코드를 main page에서 test page로 이동
